getAudioList move StorageManager and init in SoundBoosterApplication class

+++ app/src/main/java/com/vkpapps/soundbooster/utils/StorageManager.java
@@ -2,23 +2,34 @@ package com.vkpapps.soundbooster.utils;
 
 import android.app.Activity;
 import android.content.Context;
+import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.media.MediaMetadataRetriever;
+import android.net.Uri;
 import android.os.Environment;
+import android.provider.MediaStore;
+
+import com.vkpapps.soundbooster.analitics.Logger;
+import com.vkpapps.soundbooster.model.AudioModel;
 
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 
 import static android.content.Context.MODE_PRIVATE;
-/*
+
+/**
  * @author VIJAY PATIDAR
- * */
+ */
 
 public class StorageManager {
     private Context context;
+    private static List<AudioModel> audioModels = new ArrayList<>();
 
     public StorageManager(Context context) {
         this.context = context;
@@ -29,9 +40,9 @@ public class StorageManager {
         return context.getDir("userData", MODE_PRIVATE);
     }
 
-    /*
-     * @Return directory of thumbnails of song
-     * */
+    /**
+     * @Return File  private directory of thumbnails
+     */
     public File getImageDir() {
         return context.getDir("images", MODE_PRIVATE);
     }
@@ -41,26 +52,17 @@ public class StorageManager {
     }
 
     public File getDownloadDir() {
-
-        File sound_booster = new File(context.getExternalFilesDir(Environment.DIRECTORY_DOWNLOADS), "Sound Booster");
-        if (!sound_booster.exists()) {
-            sound_booster.mkdirs();
-        }
-        return sound_booster;
+        return context.getExternalFilesDir(Environment.DIRECTORY_DOWNLOADS);
     }
 
-    public void deleteMedia() {
+    public void deleteDir(File dir) {
         try {
-            File[] songs = getSongDir().listFiles();
-            if (songs != null) {
-                for (File s : songs) {
-                    s.delete();
-                }
-            }
-            File[] images = getImageDir().listFiles();
-            if (images != null) {
-                for (File s : images) {
-                    s.delete();
+            if (dir.isDirectory()) {
+                File[] songs = dir.listFiles();
+                if (songs != null) {
+                    for (File s : songs) {
+                        s.delete();
+                    }
                 }
             }
         } catch (Exception ignored) {
@@ -68,8 +70,8 @@ public class StorageManager {
         }
     }
 
-    public void copySong(File from, OnStorageManagerListener onStorageManagerListener) {
-        File to = new File(getSongDir(), from.getName().trim());
+    public void copySong(File from, String name, OnStorageManagerListener onStorageManagerListener) {
+        File to = new File(getSongDir(), name);
         //extract image from mp3
         try {
             android.media.MediaMetadataRetriever mmr = new MediaMetadataRetriever();
@@ -93,7 +95,7 @@ public class StorageManager {
 
     public void download(String name, OnStorageManagerListener onStorageManagerListener) {
         File file = new File(getSongDir(), name);
-        File out = new File(getDownloadDir(), name);
+        File out = new File(getDownloadDir(), name + ".mp3");
         copyFile(file, out, onStorageManagerListener);
     }
 
@@ -130,5 +132,36 @@ public class StorageManager {
 
     public interface OnStorageManagerListener {
         void onCopyComplete(File source);
+    }
+
+    public List<AudioModel> getAllAudioFromDevice() {
+        Logger.d("getAllAudioFromDevice: ");
+        if (audioModels.size() == 0) {
+
+            audioModels = new ArrayList<>();
+            Uri uri = MediaStore.Audio.Media.EXTERNAL_CONTENT_URI;
+            String[] projection = {MediaStore.Audio.AudioColumns.DATA, MediaStore.Audio.AudioColumns.TITLE, MediaStore.Audio.AudioColumns.ALBUM, MediaStore.Audio.ArtistColumns.ARTIST};
+            Cursor c = context.getContentResolver().query(uri, projection, null, null, null);
+
+            if (c != null) {
+                while (c.moveToNext()) {
+
+                    AudioModel audioModel = new AudioModel();
+                    String path = c.getString(c.getColumnIndexOrThrow(MediaStore.Audio.Media.DATA));
+                    String name = c.getString(1).trim();
+                    String album = c.getString(2);
+                    String artist = c.getString(3);
+
+                    audioModel.setName(name);
+                    audioModel.setAlbum(album);
+                    audioModel.setArtist(artist);
+                    audioModel.setPath(path);
+                    audioModels.add(audioModel);
+                    Collections.sort(audioModels, (o1, o2) -> o1.getName().compareTo(o2.getName()));
+                }
+                c.close();
+            }
+        }
+        return audioModels;
     }
 }
